requirements.txt
asyncio
aiohttp
aiofiles
bs4
pillow
dotenv